Spike tile added

+++ PlayMode.cpp
@@ -302,8 +302,8 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
 	ppu.sprites[0].index = 32;
 	ppu.sprites[0].attributes = 7;
 
-	unsigned offset1 = 1;
-	for (unsigned k = 0; k < walls_at.size(); k++)
+	size_t offset1 = 1;
+	for (size_t k = 0; k < walls_at.size(); k++)
 	{
 		ppu.sprites[k + offset1].x = int32_t(walls_at[k].x);
 		ppu.sprites[k + offset1].y = int32_t(walls_at[k].y);
@@ -311,9 +311,9 @@ void PlayMode::draw(glm::uvec2 const &drawable_size) {
 		ppu.sprites[k + offset1].attributes = 7;
 	}
 
-	unsigned offset2 = offset1 + walls_at.size();
+	size_t offset2 = offset1 + walls_at.size();
 	//some other misc sprites:
-	for (uint32_t i = offset2; i < 63; ++i) {
+	for (size_t i = offset2; i < 63; ++i) {
 		float amt = (i + 2.0f * background_fade) / 62.0f;
 		ppu.sprites[i].x = int32_t(0.5f * PPU466::ScreenWidth + std::cos( 2.0f * M_PI * amt * 5.0f + 0.01f * player_at.x) * 0.4f * PPU466::ScreenWidth);
 		ppu.sprites[i].y = int32_t(0.5f * PPU466::ScreenHeight + std::sin( 2.0f * M_PI * amt * 3.0f + 0.01f * player_at.y) * 0.4f * PPU466::ScreenWidth);
